set input text
