+++ client/src/components/right-section/RightSection.js
@@ -1,13 +1,20 @@
-import React from 'react';
+import React, { useState } from 'react';
 import styles from './RightSection.module.css';
-import api from '../../api/axiosInstance';
 
 import { Text, HStack, Button } from '@chakra-ui/react';
 
 import { ArrowForwardIcon } from '@chakra-ui/icons';
 
+const prompts = [
+  'What is the course CPSC455 about?',
+  'What is the goal of the CPSC455 course?',
+  'What technologies will be applied in the CPSC455 course?',
+];
+
 const RightSection = () => {
-  const PromptButton = () => {
+  const [inputText, setInputText] = useState('');
+
+  const PromptButton = ({ promptText }) => {
     return (
       <Button
         bg="#42434f"
@@ -17,26 +24,15 @@ const RightSection = () => {
         blockSize="auto"
         px={8}
         py={4}
+        onClick={() => setInputText(promptText)}
       >
         <Text fontSize="md" fontWeight={400}>
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
-          eiusmod tempor?
+          {promptText}
         </Text>
       </Button>
     );
   };
 
-  const callApi = () => {
-    api
-      .post('/users')
-      .then(response => {
-        const data = response.data;
-        alert(JSON.stringify(data));
-      })
-      .catch(error => {
-        console.error(error);
-      });
-  };
   return (
     <div className={styles.container}>
       <div className={styles.mainPanel}>
@@ -44,17 +40,22 @@ const RightSection = () => {
           CourseGPT
         </Text>
         <HStack mt={24} spacing="16px">
-          <PromptButton />
-          <PromptButton />
-          <PromptButton />
+          {prompts.map(prompt => (
+            <PromptButton promptText={prompt} />
+          ))}
         </HStack>
       </div>
       <div className={styles.inputSection}>
         <div className={styles.inputArea}>
-          <input className={styles.input} placeholder="Enter a prompt..." />
-          <button className={styles.sendBtn} onClick={callApi}>
+          <input
+            className={styles.input}
+            placeholder="Enter a prompt..."
+            value={inputText}
+            onChange={e => setInputText(e.target.value)}
+          />
+          <button className={styles.sendBtn} onClick={() => setInputText('')}>
             <ArrowForwardIcon />
-          </button>{' '}
+          </button>
         </div>
       </div>
     </div>
